feat: 랭킹 1, 2 3위 순위를 메달(medal) 이미지로 대체
